pkp/pkp-lib#1821 Clean up username strtolower calls
