+++ classes/form/validation/FormValidatorUsername.inc.php
@@ -1,22 +1,22 @@
 <?php
 
 /**
- * @file classes/form/validation/FormValidatorAlphaNum.inc.php
+ * @file classes/form/validation/FormValidatorUsername.inc.php
  *
  * Copyright (c) 2014-2016 Simon Fraser University Library
  * Copyright (c) 2000-2016 John Willinsky
  * Distributed under the GNU GPL v2. For full terms see the file docs/COPYING.
  *
- * @class FormValidatorAlphaNum
+ * @class FormValidatorUsername
  * @ingroup form_validation
  * @see FormValidator
  *
- * @brief Form validation check for alphanumeric (plus interior dash/underscore) characters only.
+ * @brief Form validation check for usernames (lowercase alphanumeric with interior dash/underscore
  */
 
 import('lib.pkp.classes.form.validation.FormValidator');
 
-class FormValidatorAlphaNum extends FormValidator {
+class FormValidatorUsername extends FormValidator {
 	/**
 	 * Constructor.
 	 * @param $form Form the associated form
@@ -24,10 +24,12 @@ class FormValidatorAlphaNum extends FormValidator {
 	 * @param $type string the type of check, either "required" or "optional"
 	 * @param $message string the error message for validation failures (i18n key)
 	 */
-	function FormValidatorAlphaNum(&$form, $field, $type, $message) {
+	function FormValidatorUsername(&$form, $field, $type, $message) {
 		import('lib.pkp.classes.validation.ValidatorRegExp');
-		$validator = new ValidatorRegExp('/^[A-Z0-9]+([\-_][A-Z0-9]+)*$/i');
-		parent::FormValidator($form, $field, $type, $message, $validator);
+		parent::FormValidator(
+			$form, $field, $type, $message,
+			new ValidatorRegExp('/^[a-z0-9]+([\-_][a-z0-9]+)*$/')
+		);
 	}
 }
